# docs/swagger.yaml
basePath: /
definitions:
  model.Post:
    properties:
      content:
        type: string
      created_date:
        type: string
      id:
        type: string
      title:
        type: string
    required:
    - content
    - title
    type: object
host: localhost:8082
info:
  contact: {}
  description: This is a blog posts server
  termsOfService: http://swagger.io/terms/
  title: Blog Platform API
  version: "1.0"
paths:
  /healthcheck:
    get:
      consumes:
      - application/json
      description: Check if the server is running and MongoDB connection is established
      produces:
      - application/json
      responses:
        "200":
          description: Everything is fine!
          schema:
            type: string
        "500":
          description: MongoDB connection failed!
          schema:
            type: string
      summary: Check server health status
      tags:
      - health
  /posts:
    get:
      consumes:
      - application/json
      description: Get all posts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Post'
            type: array
      summary: Retrieve all posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create a new post
      parameters:
      - description: Post object that needs to be added
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Post'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new post
      tags:
      - posts
  /posts/{id}:
    get:
      consumes:
      - application/json
      description: Get a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Post'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Retrieve a post by ID
      tags:
      - posts
schemes:
- http
swagger: "2.0"